Add aarch64-darwin binary

+++ pkgs/age-plugin-se.nix
@@ -3,6 +3,7 @@
   lib,
   stdenv,
   fetchurl,
+  unzip,
   runCommand,
   testers,
   age,
@@ -10,6 +11,10 @@
 let
   version = "0.1.4";
   sources = {
+    "aarch64-darwin" = fetchurl {
+      url = "https://github.com/remko/age-plugin-se/releases/download/v0.1.4/age-plugin-se-v0.1.4-macos.zip";
+      sha256 = "0i1yd5zpyw1ziirp4cr5jrlhw6xldvjrpxk9lsiay8153pdp0ksh";
+    };
     "aarch64-linux" = fetchurl {
       url = "https://github.com/remko/age-plugin-se/releases/download/v0.1.4/age-plugin-se-v0.1.4-aarch64-linux.tgz";
       sha256 = "11hyqakcp3jvr092z61b9sv1bpn088f68bxmigja7w4mb1bar6j5";
@@ -25,9 +30,22 @@ stdenv.mkDerivation (finalAttrs: {
   inherit version;
   src = sources.${system};
 
+  nativeBuildInputs = [
+    (lib.optional (system == "aarch64-darwin") unzip)
+  ];
+  sourceRoot = ".";
+
   installPhase = ''
-    mkdir -p $out/bin
-    cp usr/bin/age-plugin-se $out/bin/
+    if [ -f age-plugin-se ]; then
+      mkdir -p $out/bin
+      cp age-plugin-se $out/bin/
+    elif [ -f usr/bin/age-plugin-se ]; then
+      mkdir -p $out/bin
+      cp usr/bin/age-plugin-se $out/bin/
+    else
+      echo "age-plugin-se not found"
+      exit 1
+    fi
   '';
 
   meta = with lib; {
@@ -37,6 +55,7 @@ stdenv.mkDerivation (finalAttrs: {
     sourceProvenance = [ lib.sourceTypes.binaryNativeCode ];
     mainPackage = "age-plugin-se";
     platforms = [
+      "aarch64-darwin"
       "aarch64-linux"
       "x86_64-linux"
     ];
